Address review: lift runGC to SparkFunSuite, add retryOnOOM helper
